Add basic example test and got it running.
I did this so that I would have somewhere to start from in terms of a working
build and initial test scenario. I also played with different styles of tests
to find the one that best matched Ruby RSpec style. Beyond that, I ran nito a
couple hickups that I had to work through to get things working. For example,
I had to specify the latest version of scala in the build.sbt otherwise sbt
would try and build it against an older version of scala which caused problems
for the scalatest library.